[#516] - Changing the query to be unique for a meal id and person id, introduced slug, query the favorite meals based on person and it is restricted for user, while admin and meal designer could query any user favorite meals
